+++ swift-base/swift-base-io/src/main/java/com/fr/swift/cube/io/impl/fineio/output/ByteFineIoWriter.java
@@ -35,9 +35,9 @@ public class ByteFineIoWriter extends BaseFineIoWriter<ByteBuf> implements ByteW
     @Override
     public void put(long pos, byte val) {
         if (isOverwrite) {
-            FineIOAccessor.INSTANCE.put(writeFile, (int) pos, val);
+            FineIOAccessor.INSTANCE.put(writeFile, pos, val);
         } else {
-            FineIOAccessor.INSTANCE.put(appendFile, (int) pos, val);
+            FineIOAccessor.INSTANCE.put(appendFile, pos, val);
         }
     }
 }
